sourcecode/handler/requirements.txt
APScheduler==3.10.1
boto3==1.26.116
botocore==1.29.116
Flask==2.2.3
slack_sdk>=3.0
Jinja2==3.1.2
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
python-dotenv~=1.0.0
twilio